+++ src/directives/gen_tar_gz.rs
@@ -51,8 +51,8 @@ impl CliDirective for GenTarGzDirective {
             Some(partition_depth) => {
                 // Create path generator
                 let path_generator = crate::storage::PartitionPathGenerator::new(  
-                    note_set.len() as f32,
-                    melody_length as i32,
+                    note_set.len() as u32,
+                    melody_length,
                     self.partition_args.max_files.into(),
                     partition_depth,
                 ).unwrap_or_else(|err| {
